r/heartbeat_manager: fix register/unregister race
this patch removes two race conditions present with heartbeat manager
and registration of raft groups with the heartbeat manager.

the first race is that locks_for_range iterates over the container using
seastar::map_reduce whish is not run in the context of a single
continuation is thus the iterators for a flat_set are subject to
invalidation after synchronous invocations of register/unregister group.

The second race occurs as follows:

   1. locks_for_range: locks registered groups
[ scheduling event: newly registered group ]
     2. requests_for_range
     3. send_heartbeats

Because starting with (2) requests_for_range re-iterate over the
container of registered groups, a newly added group between (1) and (2)
will be operated upon without having been locked in the previous step.

Both of these issues is resolved by introducing a mutex to protect
mutations and iterations.

Signed-off-by: Noah Watkins <noah@vectorized.io>