storage: unify segment offset tracking
This patch moves offset tracking out of the segment_reader
and unifies it with the partial information it already had in
the segment.h itself. The current implementation is a historical
artifact when we had 2 file streams (reads,writes).

This patch moves us a step closer into having a *single* file
for reading and writing data.

All unit tests pass

Signed-off-by: Alexander Gallego <alex@vectorized.io>
Message-Id: <20200427224150.466373-1-alex@vectorized.io>