c/controller: added topic deletion feature
Topic deletion works analogical to the creation. The deletion request is
replicated via raft-0 and then applied to controller state through raft
apply notifications upcall. When deletion record is processed by the
controller logic it executes the following steps:

 1) Removes topic from metedata cache
 2) For each topic partition:
    a) removes partition mappings (NTP & group_id) from shard table
    b) removes partition via partition_manager API
    c) if current controller is leader updates partition allocator

Steps a) & b) includes updating caches on all cores as the cache
copy exists on every core.

The controller mutex is held for the whole request processing until
replicating state to raft0. Then it is hold again for processing the
raft0 append entries notification upcall. This guarantees that state
request processing and state update are atomic.

After those actions are done it returns vector of results

Signed-off-by: Michal Maslanka <michal@vectorized.io>