raft: create a raft group manager
This patch splits cluster::partition_manager into two parts: a
raft::group_manager for raft specific management tasks, and
cluster::partition_manager which retains the non-raft specific
components. This split reenforces a separation between raft and
higher-level abstractions.

This separation is important for two reasons. First, we'll likely have a
need for raft without the context of ntp-based identification.
Maintaining as much of a separtion now will make the process easier
later on.

Second, the separation of conerns will make it easier to grow the
partition manager. Currently the partition manager's function is
effectively limited to maintaining a mapping between a raft group id and
an ntp. Not only will this mapping soon be dynamic as we add migration
and balancing, but the mapping will become more complex as we make ntp's
be a further abstraction on top of unique integer identifers.

Signed-off-by: Noah Watkins <noah@vectorized.io>