r/service: fixed handling not yet registered groups
When redpanda starts it may happen that the heartbeat request will be
received by the node that does not yet have the raft group initialized.
This caused the shard_table lookup failure. Fixed the issue by correctly
handling heartbeat requests for groups that does not yet exists on
current node.

hotfix for: [ch438]

Signed-off-by: Michal Maslanka <michal@vectorized.io>