c/replicate_stm: fixed replicate stm stop condition
Replicate STM always have to wait for committed index to be greater than
or equal to local entry appended offset. If it wouldn't the it would be
unable to reliably determine what is the replication result. The entry
is required to be committed to determine the result as committed entries
can not be truncated.

Signed-off-by: Michal Maslanka <michal@vectorized.io>