storage:compaction: unify all forms of compaction under one method
This change adds a wrapper type in the log manager called
`log_housekeeping_meta.h` which effectively keeps track of the last
compaction. For now, we simply forward all of the calls to the old
`garbage_collect` methods but add in the infrastructure so
that future changes will be at the individual log level.