r/consensus: handling recovery of empty nodes
When leader have an old match index stored for the follower this may
lead to the situation in which node recovery will be impossible.
Implemented solution to this problem by checking if leader match_index
information is not up to date and triggering recovery in this case.

Note on differences from Raft paper:

The main difference between redpanda raft protocol implementation and
implementation described in the paper is the fact that in redpanda
requests reordering are first class citizens. By design Raft algorithm
is resilient for reordering but when requests are reordered the
performance penalty occur (not necessary truncations may lead to
redelivering big chunks of the logs).

Why redpanda Raft protocol requests are reordered ?

1) When not using lock our RPC layer does not guarantee that requests
   are sent in the client method call order (RPC stack may reorder them)
2) Heartbeat manager works separately from consensus without shared
   concurrency control.

How redpanda prevents request reordering on the leader ?

On the leader we track order of requests sent to the follower using its
state kept in `follower_index_metadata` structure the index is injected
into continuation handling the response when last received request is
greater than the one bound with currently processed reply it means that
the requests were reordered. This reply is discarded and does not
influence follower state.

How redpanda prevents reordering on the followers ?

On the follower redpanda doesn't truncate log when processing hearbeat
requests. This is the main difference from the original Raft algorithm.
Not truncating the log on heartbeat requests makes reordered heartbeat
'harmless' as they will not cause unnecessary redelivery of already
persisted log entries.

How the reordering prevention influence recovery mechanism ?

In original Raft when leader receives failed append entries response it
moves the follower next_index backward by one. Then all consecutive
append entries requests (including heartbeats) to this follower
are sent with the next_index-1 as a prev_log_index. When the follower
eventually return success in append_entries_reply leaders starts to send
missing batches to the follower, starting from previously established
next_index. If follower has some batches that are conflicting with the
leader log they get truncated. In this case the recovery starts when
follower returned successful response. In redpanda recovery starts when
follower returned an error and the whole process of establishing
matching leader and follower index (moving log backward) happens inside
the recovery STM. In order to make it more similar to original
implementation we would have to allow followers to truncate the log on
heartbeats which would incur big performance penalty.

Signed-off-by: Michal Maslanka <michal@vectorized.io>