log_manager: expose opening a directory of segments
This patch exposes methods used for managing a directory of segments as
well as the helper for building new segments.

This patch should be followed up (or replaced) with a better solution
that operates at a more convenient level of abstraction. As it stands
having to create the full log manager instance is overkill.

Signed-off-by: Noah Watkins <noah@vectorized.io>