c/controller: waiting for leader to be elected when creating topics
In order to make sure that redpanda is ready to handle the produce
workload right after create topic requests return, implemented waiting
for leader to be elected after topic was successfully created. By this
we are sure that all consecutive calls to metadata API will return
correct NTP leader ids.

Signed-off-by: Michal Maslanka <michal@vectorized.io>