c/controller: setting leader upfront for single node partitions
When creating partitions with only one node in replica set we can safely
assume that this node will become the replica set leader. This will
speed up the topic creation process and will allow clients to make
progress. If client will try to produce to the topic that actually
doesn't have a leader it will refresh the metadata and continue

Signed-off-by: Michal Maslanka <michal@vectorized.io>