k/groups: fix error code for sarama client
When calling Consume on a consumer group client the Kafka broker is
returning the error COORDINATOR_LOAD_IN_PROGRESS = 14 when joining the
group, but Sarama does not appear to implement the back-off / retry
logic for this error condition. It only does back-off / retry for
NOT_COORDINATOR = 16 (see here:
https://github.com/Shopify/sarama/blob/master/consumer_group.go#L227).

This case is handled in kafka java reference client (see:
https://github.com/apache/kafka/blob/trunk/clients/src/main/java/org/apache/kafka/clients/consumer/internals/AbstractCoordinator.java#L609).

Generally this error COORDINATOR_LOAD_IN_PROGRESS and NOT_COORDINATOR
and COORDINATOR_NOT_AVAILABLE are all valid errors from any of the group
membership APIs.

Sarama handles the load in progress condition for fetching offsets, but
when we experience this error via join group, the client doesn't retry
as expected.

This patch introduces a temporary fix that returns NOT_COORDINATOR which
causes the client to a refresh and retry.

Tracking here https://github.com/Shopify/sarama/issues/1715

Signed-off-by: Noah Watkins <noah@vectorized.io>