consensus: convert voted for storage to kvstore
Switches the voted for storage from being file-based to using the
key-value storage service. On boot-up the old file based storage is
read if no voted for data is found in the key-value store. Updates to
voted for are always written into the key-value store. On bootup if we
find that the key-value store contains the voted for information, the
old voted for data is removed.

The components of the old file-based voted for used for writing are
removed from the public api and used exclusively in the tests that rely
on them to create on-disk data for testing reads.

Signed-off-by: Noah Watkins <noah@vectorized.io>