storage: read past empty log segments
Prior to this change a log reader was closed if a segment read returned
no records. However, if the segment being read contained only batches
that were to be skipped (e.g. control batches) then it would appear as
though it is empty. In this case we should try to read the next segment.

Signed-off-by: Noah Watkins <noah@vectorized.io>