kvstore: check for voted for file before removing
Removing the voted for file without first checking if it exists creates
an exceptional future (ENOENT) which we catch in a handle_exception
case. This exception is fine because we would have migrated to the
kvstore management for voted for data. However, at TRACE log level the
exception is logged resulting in the exception trace and other exception
context being printed to the log.

Instead, we first check if the voted for file exists before removing to
avoid this noisy log case.

Signed-off-by: Noah Watkins <noah@vectorized.io>