Add 'vtools dbuild toolchain'
Add a new command that builds a docker image containing the vectorized/v
toolchain. This process takes into accounting scenarios where this image
is being built on hosts with cold docker cache, so that it avoids
re-building an image that has been uploaded to the docker registry that
holds these images (gcr.io currently).

In order to keep track of which images have already been built and don't
need to be re-created, it hashes dependencies of the distinct toolchain
layers (vtools, cmake, clang, go modules, etc.). For each layer, an
explicit list of file dependencies is declared, and this is used to
produce a SHA256 hash (first 10 characters) that is used to tag the
layer (docker image) so that they can be retrieved later in hosts with a
cold docker cache.

Utility functions have been placed in tools/vtools/vlib/docker.