rpc: add virtual node shard oversubscription (of 3) for transports
On a recent test on i3en.metal, we saw the throughput limited to
half of the advertised network throughput of 6GB/s. The reason was
requests for a multitude of raft groups were bottlenecked on the
lock + x-core transfer cost of data.

This patch computes the number of vnodes by reducing all shard-id's to
a known universe of 3 primes (3-vnodes) we are specifically designed for
hashing. We then use this seed to boost::hash_combine it with the node
id for an even distribution (jump_consistent_hash) among all of the cores
giving us the desired distribution of 3-vnodes * node_id