r/heartbeat_manager: added self heartbeat reply processing
Added processing of self heartbeat reply. This is important as the
protocol makes progress using append entries requests, and heartbeat are
the special case of those requests. Processing heartbeat at the current
node will make it possible to update protocol state, in particular the
commit index when there is only single node in raft group.

Fixes: [ch720]

Signed-off-by: Michal Maslanka <michal@vectorized.io>