storage:compaction: create compaction reducers
A reducer returns a bitmap at the end_of_stream() function.

We move the old first_pass* reducer into a file that combines
the logic in a testable unit.

This unit contains reducers to reduce offsets to natural indexes
as described in the design document as well as keys to model::offsets to
natural indexes.

See section 4 of the design doc
https://docs.google.com/document/d/10iRzQZcaHSatS17xmey9FJY5nqkWZVmxF_7X9ufzf_4/edit#heading=h.mvlqwsjv79vf