storage:segment_utils: bugfix ss::open_flags manipulation
while creating segments we do not need to check if the file
has been created, we can simply pass ` x | ss::open_flags::create`
to the open call.

In case of `x | ss::open_flags::excluxive` checking if the handle
is already open is a race condition and in fact we should throw.