r/mux_state_machine: implemented multiplexing stm on top of state_machine
The multiplexing STM allows building multiple state machines on top of
single consensus instance. The mux_state_machine dispatches state
applications to correct state implementations using
is_batch_applicable method result as a discriminator. Each state
implementation must provide the `apply_update` method that is called by
the state machine when applying entries. The `apply_update` method
returns an error code that is propagated to the replicate caller.
The dispatching STM is designed in such way to allow state updates
without locking. The result is determined when applying update to the
state. Thanks to this approach it is easy to implement the optimistic
locking concurrency control in state.

Signed-off-by: Michal Maslanka <michal@vectorized.io>