c/partition_allocator: added allocation units
The optimistic locking that is going to be used in topics state requires
the allocation to be two step process. In the first step before
replication the topic_frontend will reserve allocation units and then
when applying the topic table update dispatcher will confirm those
reserved allocation units. When recovering the allocation units are
updated in a single step.

Signed-off-by: Michal Maslanka <michal@vectorized.io>