c/commands: implemented controller commands
Controller state updates are represented in terms of commands. Each
command represent different type of action that is going to be executed
over state when applying replicated entry. The controller_command is a
base type that underlies all commands used in the controller the
type provides information required for dispatching updates in
`raft::mux_state_machine` and automatic serialization/deserialization
handling.

Signed-off-by: Michal Maslanka <michal@vectorized.io>