rpk/tuners: Add fstrim tuner
Add a tuner to ensure that a weekly fstrim systemd timer is active.

Most of the latest distros (Debian >=10, Ubuntu >=18.04, RHEL >=7,
Fedora >=31) already ship with a standard systemd timer and service
to run fstrim weekly, so in most cases the tuner's work will be to
just start it. However, for older distros that use systemd, or
custom images where the timer was removed, the tuner will also
install and start a timer and a service.