storage:compaction: eliminate additional user api of load_footer()
In order to read the compacted index, we need to eliminate the footer
size. However, because of multiple versions, we need to first read
the actual footer which may have versions switches and therefore different
sizes on disk.

Remove the need for the caller to specify reading the footer explicitly
and more clearly express the intent of the caller by transitively
hydrating the datastructure in case it's not loaded.