quota_manager: apply a configurable max delay
Along with the commit:

  commit 13767b5224f7a195f447bedbf308d00bc29fa713
  Author: Noah Watkins <noah@vectorized.io>
  Date:   Mon Jun 8 11:38:17 2020 -0700

      kafka: generate default throttle value

      Sometimes a throttle value on a request response is not set by accident,
      resulting in a very large value being sent back to the client. This sets
      a default throttle value of zero on the generated request structure.

this fixes [ch398] which seeks to address errant throttle values,
erroring on the side of disabling throttling. follow up work may address
verifying that throttling is always set on every request. this will
require more effort, as there is not standard--it is on a per request
basis. we can probably use code gen to help with this.

Signed-off-by: Noah Watkins <noah@vectorized.io>