storage: remove segment management from log manager
This is primarily a code movement patch.

This patch move segment management out of the log manager. Segment
opening and creation are associated with the segment type, and segment
recovery is associated with the segment_set type.

The few pieces of state that were queried from the log manager instance
are passed into the new routines as parameters.

In addition to being rather irrelevant to the log manager, moving the
segment management out of the log manager is needed to clean-up the
key-value store interface.

fixes: [ch652]

Signed-off-by: Noah Watkins <noah@vectorized.io>