storage: register kvstore metrics on startup
Registering them in the constructor makes disabling them in tests
difficult because there needs to be an active component run to disable
metrics through thread local configuration _before_ the key-value store
is constructed. registering the metrics in start() makes this trivial;
we can always run the code to disable metrics before calling start().

Signed-off-by: Noah Watkins <noah@vectorized.io>