storage: implement precise prefix truncation
The basic strategy is to remove all segments below the new starting
offset. if the result is an empty log then the next segment created will
begin at the new starting offset.

Otherwise, if the new starting offset lands in the middle of a segment
then we do not attempt to reclaim the space associated with the prefix
of the segment but instead implement log offset visibility logically
based on the starting offset.

This patch also removes from offset_stats the term associated with the
starting offset. This was unused before this patch. Prefix truncation
means that it would be valid to have a starting offset >= 0 without any
actual segments. It therefore is not clear what having a term would
actually mean.

Signed-off-by: Noah Watkins <noah@vectorized.io>