controller: fix controller delta topics backend race
the reconcile_topics method in the controller backend uses
ss::parallel_for_each to operate on topics_delta which creates futures
that each contain a reference into topics_delta.

the entire loop runs under the topics semaphore, however,
start_topics_reconciliation_loop adds elements to the topics_delta set
without acquiring the lock.

because reconcile_topics may run in parallel via housekeeping and the
reconcillation loop, the addition to topics into the set may interleave
with the parallel leading to an invalid reference being operated on if
increasing the size of the set causes a resize event.

fixes: [ch835]

Signed-off-by: Noah Watkins <noah@vectorized.io>