packaging: Patch the compiled bins with the abs path to interpreter
The binaries that are shipped with redpanda need to be patched
so that the interpreter set in the ELF headers matches the path
to the one shipped with redpanda packages (/opt/redpanda/lib/ld.so),
allowing debug symbols to be loaded by gdb when debugging.

The current way of executing the shipped bins relies on "discovering"
the interpreter at runtime, which confuses gdb into thinking ld.so
is the bin, instead of redpanda.

Update the RPM spec file to account for these changes.