kafka: add support for noop responses
This adds a flag to the kafka response type that request handlers can
use to mark a response as a noop response. A noop response is a response
that is fully processed (e.g. quota, stats, probes accounting) but is
not written to the output connection.

The primary use case is supporting acks=0 produce requests whose
responses are not expected by clients.

Signed-off-by: Noah Watkins <noah@vectorized.io>