storage: roll without forcing a write lock
Requiring a write lock to be obtain when rolling a log segment means
that a reader holding the read lock could hold up appender progress for
an arbitrary amount of time once a segment became full.

This patch introduces optimistic locking for rolling. If the write lock
for a segment can be obtained without blocking then the normal process
of releasing the appender occurs.

However, if the write lock cannot be obtained then the minimal amount of
work required to make the segment appear from the outside that it has
had its appender released is performend, and the actual work is
scheduled in a background fiber that runs when the write lock does
become available.

fixes: [ch519]

Signed-off-by: Noah Watkins <noah@vectorized.io>