vtools/config: install llvm to separate subfolder
having LLVM installed on /vectorized for CI builds is causing system
binaries such as clang++, git, etc. to pick up the sanitized version of
zlib that vtools builds as part of 3rdparty.cmake.in instead of the
system one.

By moving the llvm prefix to a subfolder, this doesn't happen anymore