storage: use generic iobuf for snapshot metadata
The original version of the snapshot feature stored snapshot metadata
specific to raft. The first real user of snapshot was key-value store
which didn't use all of the raft metadata, but made use of a single
metadata field to store an offset associated with the snapshot.

Now raft is starting to use the snapshot manager and needs to store
other data such as configuration history and timestamps. Instead of
letting the metadata schema further diverge between sub-systems that use
the snapshot manager, this patch changes the metadata to be a small
generic iobuf.

The key-value store now includes its single integer in the metadata
explicitly.

Signed-off-by: Noah Watkins <noah@vectorized.io>