compression:lz4: bugfix frame regrowth for lz4f format
When lz4f compresses really well and the kafka client does not
embed the correct size in the frame.info, then we guestimate the
size of the input*4 as the destination buffer. When payloads
are large - say 64MB - we have to regrow (at 1.75x rate) to
accomodate for the kafka drivers not embeding the correct lz4 frame
information.