r/offset_monitor: added last_notified field
When calling `raft::state_machine::wait` the logic checks if requested
offset was already applied to state machine and if it does it return
immediately. The check is based on `_next` field value. The `_next`
field is updated after all batches were applied to state machine while
`_waiters` are being updated during each batch application. This caused
race condition as `wait` was called on offset monitor for which the
offset was already applied. Introduced checking if offset was already
applied in offset_monitor. Introducing check in the monitor itself will
prevent race conditions in future as order of `_last_applied` is always
the same as order of `notify` calls.

Race condition example:

Consider this (assuming that `10 >= _next & ss::gate([]{ c1 }`) is
atomic scheduling unit

1) f1: reader.consume(batch_applicator(this))       // in `state_machine::apply()`
2) f1: _machine->apply(std::move(batch))            // in `batch_applicator::operator()`
3) f1: _waiters->_notify(10)                        //in `batch_applicator::operator()`
4) <schedule>
5) f2: ss::gate([]{ 10 >= _next; return _waiters.wait(10); } // in `wait(model::offset)`
6) <schedule>
7) f1: batch_applicator::end_of_stream()
8) f1: _next = last_applied + model::offset(1); // state_machine::apply()

_next value is checked in step 5 while it is updated in step 8 even
though apply & notify are executed in steps 2 and 3.

Signed-off-by: Michal Maslanka <michal@vectorized.io>