r/tests: fixed mux state machine recovery test
The recovery test was using log builder pipe operator to persist
batches, before starting mux state machine to test recovery. Persisting
test batches to the log was single expression. In order to track the
offset of each batch persisted to the log we used `offset` variable that
was post incremented after using it to create batches. The evaluation
order of piped expression caused the log to be corrupted as offsets were
assigned in incorrect order. Fixed the problem by splitting building log
into multiple expressions.

Signed-off-by: Michal Maslanka <michal@vectorized.io>