copro: add close, closeCoprocessorManager and executeCoprocessorOnRequest
the close method stop the server's connection, after that,
it stops the inotify (filesystem watcher) process and
returns a promise when is done.

the closeCoprocessorManager method stop the inotify process

executeCoprocessorOnRequest add a listener for "readable" socket
event, and apply applyCoprocessorToCoprocessorRequest method
for each new request.

Note: in this patch, The Serializer and Deserializer haven't
implemented yet

---
since v4:
  - fix typo