bamboo: mount v into test container context
This patch mounts the source tree as a read-only bind mount so tests can
access tree contents.

This is a solution for making useful bits from around the source tree
available to tests. For instance, helpful scripts in tools directory.
But, dockerfiles need their sources to be available in reproducible
locations, so we can't grab various dependencies from around the source
tree.

Signed-off-by: Noah Watkins <noah@vectorized.io>