c/members_manager: handling of configuration updates
Added configuration updates handling to members manager. The
configuration is updated when node starts and it establish that the
configuration changed. Configuration request is sent to one of the nodes
as the information about the leader may never be available locally (f.e.
when ip address changed). When request is received by one of the nodes
that isn't the current leader it sends is further to current leader
controller. When members manager is notified about configuration change
it updates the `cluster::members_table` with new information.

Signed-off-by: Michal Maslanka <michal@vectorized.io>