c/members_manager: added checking node configuration invariants
Added checking if node id set in redpanda configuration file changed.
If so the exception is thrown during startup causing redpanda to stop.
The check is made before cluster joining sequence so the node with
incorrect configuration wouldn't corrupt the cluster state. The node id
can only after node is removed (not yet supported) from the cluster and
all data are deleted. Then the node can rejoin the cluster with new id.
Node id is persisted during first start and it is persisted in local
kv-store.

Signed-off-by: Michal Maslanka <michal@vectorized.io>