bugfix:storage:compaction: always close segment_appender on reducer
fixes [ch1001]. Properly handle segment appender closing so we
do not obscure the actual exception when we detect data corruption.

Added exception information to be able to debug future crashes.
The bugfix is that we cannot rely on `end_stream` function call
to be call if we have an exception during processing.