r/raft_group_fixture: handling prefix truncated log validation
Log replicated by raft can be prefix truncated independently on each of
the nodes. This situation is correct and should be accounted for when
validation log replication in raft tests. Added handling of prefix
truncated logs to validation logic. For validation max start offset is
established and all following batches are expected to be exactly the
same on all raft group members.

Signed-off-by: Michal Maslanka <michal@vectorized.io>