tests: copy ducker utility from kafka with mods
ducker is a utilty designed to operate on a group of docker containers.
for the most part this is analgous to what bamboo is currently which a
tool for spinning up docker containers and setting up networking. the
main difference is that ducker and ducktape assume that docker
containers are not running any services. they are treated more like
nodes and operations are explicitly executed within them, in contrast to
starting a container assuming it runs redpanda.

It's modified for redpanda, mostly to remove a few kafka specific bits:

- Removed kafka branding and kafka-specific variables
- Removed assumption about java base docker image
- Changed container mount paths for v dev tree

Signed-off-by: Noah Watkins <noah@vectorized.io>