r/configuration_manager: introduced configuration manager
Introduced the `raft:configuration_manager` class. The class is
responsible for keeping all reachable raft group configurations
together with their offsets. The configuration manager stores all
configurations in storage::kvstore, to speed up raft groups recovery.
Manager updates the last_highest knows offset every 64MB of persisted
data and when configuration is added to the manager. Thanks to this
when raft group is starting it only has to read up to 64MB of
data to find configurations that may not be included in the
configuration manager.

Signed-off-by: Michal Maslanka <michal@vectorized.io>