storage: add more control over max bytes in log reader
Before this change the reader config's max bytes was only enforced after
it was exceeded.  The new default behavior is to only allow max bytes to
be exceeded for the first batch read so that progress will be made.

This patch also adds a strict_max_bytes flag that disallows max bytes to
be exceeded even in the case of the first batch. This is useful for
kafka fetch request which will set this flag after any batch has been
read so that it in turn doesn't exceed its limits.

Signed-off-by: Noah Watkins <noah@vectorized.io>