storage: do not restrict the full segment reader bytes
The full segment reader is created to read a range of offsets, but also
introduces a restriction on the max number of bytes to be the size of
the segment. Since the reader is consumed in a one-shot fashion, the
user of the reader is sensitive to any sort of disconnect between the
tracked size of the segment and the exact byte consumption accounting in
the log reader.

We saw this with 2d37228 in which an off by one calculation caused the
full segment to not be processed. This patch removes that disconnect by
relying only on the offset ranges for the reader limits.

Signed-off-by: Noah Watkins <noah@vectorized.io>