r/consensus: do nothing when the same configuration is added twice
It may happen that the configuration batch is the last batch of
snapshot. In this case consensus adds the configuration to the prefix
manager although it already exists. We do not have to throw an exception
in this case as long as the added configuration is equal to the one that
is already stored in configuration manager at the same offset.

Signed-off-by: Michal Maslanka <michal@vectorized.io>