Adds a tool for checking linearizability
Linearizability [1] is a property of a concurrent / distributed system. We say
that an access to data structure is linearizable if even when operations
overlap, each operation appears to take place instantaneously.
This means that we can think of a linearizable data structure as if it was
implemented as a remote single threaded server.

The algorith is inspired by the "Testing shared memories" by Phillip B. Gibbons
and Ephraim Korach .

We use write_id to do read-mapping and use compare and set on write_id to
resolve write-order. Since we require that operations are already sorted the
complexity falls form O(N ln N) class to O(N) class where N is the length of
a history. It looks like shifting of the responsibility but when all
operations are initiated from the same node they already happen in time order
so we don't need to sort and can check consistency on the fly.

[1] https://en.wikipedia.org/wiki/Linearizability
[2] http://citeseerx.ist.psu.edu/viewdoc/summary?doi=10.1.1.107.3013