r/log_eviction_stm: passing in abort_source reference
Passing in an `abort_source` reference as a `log_eviction_stm`
constructor parameter to be able to interrupt monitoring log evictions.

Log eviction uses `raft::monitor_log_eviction` method to be notified
about segment evictions that are going to be executed by storage layer.
When stopping, monitoring log eviction has to be interrupted.

Signed-off-by: Michal Maslanka <michal@vectorized.io>