log_manager: prevent race between dir removal and dir listing
If directory removal operation races with directly listing then
internal to seastar a syscall failure is ignored and the stream is
released without closing, leading to a broken promise. This serializes
the checking and removal of empty topic directories so they do not race
with parallel ntp removals originating from the controller backend.

See more details in:

https://app.clubhouse.io/vectorized/story/1104/broken-promise-in-directory-walker

fixes: [ch1104]

Signed-off-by: Noah Watkins <noah@vectorized.io>