s/log: using max_collectible_offset to control eviction progress
Introduced max_collectible_offset to control eviction progress. The
max_collectible_offset will be updated by raft basing on its commit
index and snapshots.

Eviction of log segments has to be synchronized with taking a snapshot.
(i.e. the snapshot has to be taken before the eviction for the same
offset will happen). Previous approach was using the segment read write
lock to synchronize evictions. The read lock units were passed to the
eviction monitor, while the eviction of segments happened under the
write lock. This lead to the situation in which Raft couldn't perform
any operations (flush, read) on the log segment that were going to be
evicted as those operations are executed under the read lock. Read lock
couldn't been acquired as waits on seastar rwlock are released in FIFO
order. Waiting for write lock prevents all following waits for read lock
from being completed. The described situation led to deadlock where
eviction process was waiting for release of read lock, which can only be
released when commit_index of raft will be greater than dirty_offset of
segment that is going to be evicted. The commit_index couldn't however
be updated as reading batches required to recover the follower wasn't
possible because eviction process was waiting for write lock so lock
necessary to read batches for eviction couldn't be acquired.

Signed-off-by: Michal Maslanka <michal@vectorized.io>