Improves performance of kvrsm by removing mutex
We use a map indexed by offsets to keep track of ongoing requests.
The offset becomes known only after consensus->replicate returns.
The problem is a request may be applied before the replicate is
done so we can't set a hook for an already executed request.

The problem is solved with  mutex but it has 20x performance penalty.
This patch introduces a surrogate id and initializes the map before
calling consensus->replicate eliminating the need for the mutex.