s/disk_log_impl: fixed race between prefix truncate and eviction
Incorrect order of actions leading to updating the start offset after
log eviction caused the race condition between eviction and prefix
truncation. This caused consensus to take snapshot with last included
offset that was smaller than current log start offset which in the end
lead to triggering an assertion as term is not available for offsets
lower than the log start offset.

Signed-off-by: Michal Maslanka <michal@vectorized.io>