r/logger: added ntp to raft context log
Added an NTP instead of node id to raft context logger prefix.

The raft context logger add prefix to each log line logged in context of
raft group. It is very helpful to have the NTP in the log context as it
allows to immediately correlate raft & storage logs. Node id information
on the other hand is redundant as user already know from where the logs
come from.

Signed-off-by: Michal Maslanka <michal@vectorized.io>