s/segment_set: fixed recovery when last segment is empty
Recovering previous segment if last one was empty and got deleted.

When last segment is empty the previous one is not being recovered, even
tough its state may be invalid. Added checking if segments that are
going to be recovered are empty before entering the recovery loop. When
last segment is empty we will try to recover previous one.

Signed-off-by: Michal Maslanka <michal@vectorized.io>