storage:compaction:bugfix: fix bad optional access
calling operator-> on an optional with a default constructable
value will return the default ctor which is invalid in the context
of trying to decide if we have reached the end of the stream.

The result of this patch is that we wouldn't read compacted indexes
as required during rebuild.