bugfix: always flush the segment before closing
When we don't have an appender, this will be a noop.
However, it is crucial that we flush before closing so that
the `offsets()` metadata gets updated corectly on segment closing.

This is only a problem for tests, as we never close segments
while the system is running. It is however a bug when we
decide to implement lazy opening.