build:update seastar build
We haven't pulled seastar in over 6 months. Since
then seastar updates to API3 and deprecated API2 (what we are on)
in addition deleted labels (grafana API generation might not work)
and added noexcept assertions to abort source.

Last, it now requires valgrind headers for debug. I created
a separate repo for this and added it as a dependency.