Updates gobekli to log availability data
The number of successful operations per last second taken each second gives
an insight about availability of a system. For example when it drops to zero
the system is unavailable.

The change makes gobekli not only to print availability log to the console but
also log it to a file.